Makefile Working. Started Python GUI. May need to alter C functions

+++ dll/includes/dataprocessing.h
@@ -1,7 +1,7 @@
 #ifndef __DATAPROCESSING_H
 #define __DATAPROCESSING_H
 
-double mean(double num[], int n);
+float mean(float num[], int n);
 double stdev(double num[], double mean, int n);
 double rms(double num[], int n);
 double Amp(double num[], int n);
